requirements.txt
asgiref==3.4.1
Django==3.2.7
django-dbbackup==3.3.0
django-filter==2.4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
